backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
pymysql==1.1.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
nltk==3.8.1
Werkzeug==3.0.1
email-validator==2.1.0
